Move onMapPress functionality to MapContext
